fix `mlflow` param not fitting into `sqlalchemy` max length
Note that this is not an internal limitation in `mlflow` or `sqlalchemy`. It can be fixed by increasing the limitation from 500 chars to longer. Yet, it seems we would be better of with logging these long info which ARE NOT PARAMS in fact, as a raw string log rather than an model param (since we don't even use it to compare models, it is just there as feature engineering reference)