make `potential` aware of dynamic payload variables
Previously, `potential` was computed for a static list of variables. Now, it is only computed for the variables provided in the payload. The main benefit is that since initial questions have the highest potential, then the user will understand why filling these questions (payload variables) is important, and more importantly, how much importance they have.