add ability to not set the args for `AutoML` fit
Previously, you had to provide a value if it is defined in the configs. But, since the default value in `flaml` is designed in a way that not providing that parameter does something different rather than setting a default value (e.g., `max_iter=0` is not the same as not providing `max_iter` at all), hence, now, by providing emptry string `""` for the args in the config, you can unset.

This is used on `max_iter` here to use the heuristics of `AutoML` for hyperparameter tuning based on the value of `time_budget`.