add responses and importances for *bank balance*
in simple terms, we have a dictionary where keys are the string name of possible responses from the user which are actually the upper and lower bound for a mapping of continuous value to this range. The values, are out domain expert knowledge of the importance of that response, analogous to XAI values.