add continuous form of manual param builder
so far, we only had the ability to create manual parameters `ParameterBuilderBase` for only and only *categorical* variables. Now, a simple extension on top of `ParameterBuilderBase` enables us to have an arbitrary *interpolator* for obtaining the *importance* values for that continuous param. Any new continuous param should extend `ContinuousParameterBuilderBase` class, and as before, implement the `_modifier` methods. The bounds or any other constant needed for this new continuous variable needs to be defined inside the `constant` module. Also, the interpolator for the importance values needs to be defined inside `interpolator` module.