add ability to use a human-readable description for params
when reporting XAI values, the name of the parameter and chosen response is concatenated and reported alongside the XAI value (or importance in case of manual params). In some cases, these values are not human readable and needs to be replaced by a descriptive one. Now attribute `description` is doing that. For domain-specific manual params, these human-readable descriptions have been provided yet can be overriden on-the-fly by the user.